fix: split init_sturct and add header files
