+++ includes/error.h
@@ -1,0 +1,39 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   error.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: haryu <marvin@42.fr>                       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/08 14:03:14 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/08 14:10:07 by haryu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef ERROR_H
+# define ERROR_H
+
+/* error macro */
+/* Single quotes */
+# define SQ -39
+/* Double quotes */
+# define DQ -34
+/* Both quotes */
+# define BQ -183
+/* Semi colon */
+# define SEMI -59
+/* Comma */
+# define COM -44
+/* Input redirection */
+# define IR -60
+/* Output redirection */
+# define OR -62
+/* Heredoc redirection */
+# define HD -120
+/* Output Append mode */
+# define OAP -124
+/* PARSE */
+# define PARSE -1
+
+
+#endif
