+++ includes/readline/keymaps.h
@@ -26,15 +26,15 @@
 extern "C" {
 #endif
 
-#if defined (READLINE_LIBRARY)
+//#if defined (READLINE_LIBRARY)
 #  include "rlstdc.h"
 #  include "chardefs.h"
 #  include "rltypedefs.h"
-#else
-#  include <readline/rlstdc.h>
-#  include <readline/chardefs.h>
-#  include <readline/rltypedefs.h>
-#endif
+//#else
+//#  include <readline/rlstdc.h>
+//#  include <readline/chardefs.h>
+//#  include <readline/rltypedefs.h>
+//#endif
 
 /* A keymap contains one entry for each key in the ASCII set.
    Each entry consists of a type and a pointer.
