+++ src/skip_arg.c
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   skip_arg.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cgim <cgim@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/10 11:29:10 by cgim              #+#    #+#             */
+/*   Updated: 2022/06/10 11:29:11 by cgim             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	skip_arg(char *cmd, int i)
+{
+	while (cmd[i] != '\0')
+	{
+		if (cmd[i] == '<' || cmd[i] == '>' || cmd[i] == ' ')
+			break ;
+		else if (cmd[i] == '"' || cmd[i] == '\'')
+			i = skip_quote(cmd, i);
+		else
+			i++;
+	}
+	return (i);
+}
