+++ src/skip_quote.c
@@ -1,0 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   skip_quote.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cgim <cgim@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/10 19:48:31 by cgim              #+#    #+#             */
+/*   Updated: 2022/06/10 19:48:33 by cgim             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	skip_quote(char *cmd, int i)
+{
+	int	q;
+
+	q = 0;
+	if (cmd[i] == '"')
+		q = 2;
+	else
+		q = 1;
+	i++;
+	while (q != 0)
+	{
+		if ((cmd[i] == '"' && q == 2) || (cmd[i] == '\'' && q == 1))
+			q = 0;
+		i++;
+	}
+	return (i);
+}
