+++ src/skip_whitespace.c
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   skip_whitespace.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cgim <cgim@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/10 11:27:34 by cgim              #+#    #+#             */
+/*   Updated: 2022/06/10 11:27:35 by cgim             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	skip_whitespace(char *cmd, int i)
+{
+	while (cmd[i] == ' ')
+		i++;
+	return (i);
+}
